requirements.txt
requests>=2.31.0
pycryptodome>=3.20.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
click>=8.1.0
faster-whisper>=1.0.0
openai>=1.0.0
